Update Register Synchronization

+++ src/main/java/org/frc5411/lib/instrument/gyroscope/archetype/PigeonGyroscope.java
@@ -152,7 +152,7 @@ public class PigeonGyroscope extends Gyroscope<Pigeon2> {
           .toArray());
         UPDATE_TIMESTAMPS.clear();
       }        
-      Article.setMeasurements(IntStream.range((0), (int) Figures.minimum(Yaw.length, Pitch.length, Roll.length)).mapToObj((Index) -> 
+      Article.setMeasurements(IntStream.range((0), Figures.minimum(Yaw.length, Pitch.length, Roll.length)).mapToObj((Index) -> 
         new Rotation3d(
           Roll[Index],
           Pitch[Index],
