+++ src/main/java/org/frc5411/lib/instrument/module/archetype/MockModule.java
@@ -174,7 +174,7 @@ public class MockModule extends Module<DCMotorSim,Optional<Object>> {
         UPDATE_TIMESTAMPS.clear();
       }
 
-      Article.setMeasurements(IntStream.range((0), (int) Figures.minimum(Translations.length, Rotations.length)).mapToObj((Index) -> 
+      Article.setMeasurements(IntStream.range((0), Figures.minimum(Translations.length, Rotations.length)).mapToObj((Index) -> 
         new SwerveModulePosition(
           (Translations[Index] - getDescriptor().TranslationalOffset) * getDescriptor().Radius, 
           Rotation2d
